add class ResultSearchByLocationTest
move testResultSearchByLocationPage
