+++ src/test/java/by/itacademy/taf/ostrovok/ui/tests/BaseTest.java
@@ -1,4 +1,4 @@
-package by.itacademy.taf.ostrovok.ui;
+package by.itacademy.taf.ostrovok.ui.tests;
 
 import by.itacademy.taf.ostrovok.ui.driver.DriverSingleton;
 import org.testng.annotations.AfterMethod;
